requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-Cors==4.0.0
greenlet==2.0.2
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
openai==0.8.0
Pillow==8.4.0
python-dotenv==0.20.0
requests==2.27.1
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
zipp==3.6.0
